SubscriptionsContext replaced by useSubscriptions
