+++ assets/js/pages/ProjectCheckInNewPage/page.tsx
@@ -4,16 +4,18 @@ import * as Pages from "@/components/Pages";
 import * as Paper from "@/components/PaperContainer";
 import * as Icons from "@tabler/icons-react";
 
-import { SubscriptionsProvider, findNotifiableProjectContributors } from "@/features/Subscriptions";
-import * as Forms from "@/features/projectCheckIns/Form";
+import { findNotifiableProjectContributors } from "@/features/Subscriptions";
+import { Form, useForm } from "@/features/projectCheckIns/Form";
 import { useLoadedData } from "./loader";
 import { Paths } from "@/routes/paths";
 import { useMe } from "@/contexts/CurrentUserContext";
 
 export function Page() {
-  const { project } = useLoadedData();
   const me = useMe()!;
+  const { project } = useLoadedData();
+
   const people = findNotifiableProjectContributors(project, me);
+  const form = useForm({ project, mode: "create", author: me, notifiablePeople: people });
 
   return (
     <Pages.Page title={["Check-In", project.name!]}>
@@ -21,21 +23,11 @@ export function Page() {
         <Navigation project={project} />
 
         <Paper.Body>
-          <SubscriptionsProvider people={people}>
-            <Form />
-          </SubscriptionsProvider>
+          <Form form={form} />
         </Paper.Body>
       </Paper.Root>
     </Pages.Page>
   );
-}
-
-function Form() {
-  const me = useMe()!;
-  const { project } = useLoadedData();
-  const form = Forms.useForm({ project, mode: "create", author: me });
-
-  return <Forms.Form form={form} />;
 }
 
 function Navigation({ project }) {
